Use QuokkADB for ADB bus capture
src/sniffer turns a QuokkADB into an ADB bus decoder.
It outputs the bus data over a UART. Currently this is SWO line on the
debug header. This works well with a Raspberry PI Pico board flashed
with the Pico probe CMSIS-DAP emulator. Pin 6 on the Pico probe can read
SWO line and output over a USB CDC serial port.
(https://github.com/raspberrypi/picoprobe).

Currently it records every ADB transaction with a millisecond timestamp
and if a talk register 0 is requested it will display both mouse
and keyboard event response. It doesn't decern which particular
device is doing the output but since it records which address the
transaction is occurring on, it should be easy enough to decern.