Reorganized the quokkadb.cpp
The board now boots without needing to set the clock speed
after the first core is loaded. Was able to put most of the
setup code in initVariant. Removed the UART setup code
as now it is done via Arduino. Printf now works too.

Currently TinyUSB sees a USB stick as a USB device but the code
specific to mass storage doesn't mount for some unknown reason.