Add German and Swiss DE/FR keyboard layouts
This adds both German and Swiss layouts. It is a naive attempt to do so,
as it translates key positions based on the original Apple ADB keyboards
and doesn't account for current Mac and Windows USB layouts.

It also might fix a missing key on the French layout that is the
accent / British pound key.