Add Ireland/UK keyboard layout
This adds layout for UK keyboards based on USB Apple keyboard layout.
The keys for Microsoft USB keyboards are slightly different.
One solution would be to add a abstraction layer for certain keys so
pressing '~' on a Microsoft keyboard sends Shift-'`'. But that would
interfere with key combinations and multi-key presses for games.